Add sample Unreal project and create URN plugin for Unreal
Integrating the URN core library into Unreal by referencing the files outside of the Unreal plugin root.
This isn't meant to be done this way, so UBT only supports external include paths, and not source paths.
So for the sake of this example, the .cpp files are included into a single translation unit inside of the Sources folder in the plugin.